renamed to node-image
