+++ pkg/cli/admin/nodeimage/OWNERS
@@ -12,3 +12,4 @@ reviewers:
   - pawanpinjarkar
   - rwsu
   - zaneb
+
